# swagger.yaml
openapi: 3.0.3
info:
  title: AI Guard - Enterprise AI API Management Platform
  description: |
    AI Guard is a comprehensive API management platform that provides secure, monitored access to multiple AI providers (OpenAI, Anthropic, Google Gemini) with advanced features like multi-tenancy, authentication, rate limiting, usage tracking, and audit logging.
    
    ## Authentication
    
    AI Guard supports two authentication methods:
    
    1. **Firebase Authentication**: Use Firebase ID tokens for user authentication
    2. **Personal Access Tokens (PAT)**: Use generated tokens for server-to-server communication
    
    ## Rate Limiting

    All API endpoints are subject to rate limiting based on user tier:
    - **Free**: 10 requests/minute
    - **Pro**: 100 requests/minute  
    - **Enterprise**: 1000 requests/minute
    
    ## Error Responses
    
    All endpoints return structured error responses with helpful suggestions:
    ```json
    {
      "error": {
        "type": "AUTHENTICATION_ERROR",
        "message": "Invalid or expired token",
        "statusCode": 401,
        "timestamp": "2024-01-01T00:00:00Z",
        "suggestions": ["Check your authorization header"]
      }
    }
    ```
  version: 2.0.0
  contact:
    name: AI Guard Support
    email: support@aiguard.dev
  license:
    name: ISC
    url: https://opensource.org/licenses/ISC
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.aiguard.dev
...

...
components:
  securitySchemes:
    FirebaseAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID Token obtained from Firebase Authentication
    PersonalAccessToken:
      type: http
      scheme: bearer
      bearerFormat: PAT
      description: Personal Access Token generated via API
    AdminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: Admin secret key for administrative operations
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        ownerId:
          type: string
        members:
          type: array
          items:
            $ref: '#/components/schemas/ProjectMember'
        memberCount:
          type: integer
        apiKeyCount:
          type: integer
        userRole:
          type: string
          enum: [owner, admin, member]
        settings:
          $ref: '#/components/schemas/ProjectSettings'
        usage:
          $ref: '#/components/schemas/UsageMetrics'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, ownerId, members, createdAt, updatedAt]
    ProjectMember:
      type: object
      properties:
        userId:
          type: string
        role:
          type: string
          enum: [owner, admin, member]
        addedAt:
          type: string
          format: date-time
      required: [userId, role, addedAt]
    ProjectSettings:
      type: object
      properties:
        rateLimitOverride:
          type: object
          properties:
            windowMs:
              type: integer
              description: Rate limit window in milliseconds
            maxRequests:
              type: integer
              description: Maximum requests per window
        quotaOverride:
          type: object
          properties:
            monthlyLimit:
              type: integer
            dailyLimit:
              type: integer
        allowedProviders:
          type: array
          items:
            type: string
            enum: [openai, anthropic, gemini]
        webhookUrl:
          type: string
          format: uri
    UsageMetrics:
      type: object
      properties:
        currentMonth:
          type: object
          properties:
            requests:
              type: integer
            tokens:
              type: integer
            cost:
              type: number
              format: float
        currentDay:
          type: object
          properties:
            requests:
              type: integer
            tokens:
              type: integer
            cost:
              type: number
              format: float
        lastUpdated:
          type: string
          format: date-time
    CreateProjectRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Project name
      required: [name]
...
    ApiKey:
      type: object
      properties:
        keyId:
          type: string
        provider:
          type: string
          enum: [openai, anthropic, gemini]
        maskedKey:
          type: string
          description: Masked version of the API key for security
        isActive:
          type: boolean
        addedBy:
          type: string
          description: User ID who added the key
        addedAt:
          type: string
          format: date-time
      required: [keyId, provider, maskedKey, isActive, addedBy, addedAt]
    AddApiKeyRequest:
      type: object
      properties:
        provider:
          type: string
          enum: [openai, anthropic, gemini]
          description: AI provider name
        apiKey:
          type: string
          minLength: 10
          description: The actual API key from the provider
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Descriptive name for the key
      required: [provider, apiKey]

    # Usage and Analytics schemas
    UsageStats:
      type: object
      properties:
        projectId:
          type: string
        period:
          type: object
          properties:
            start:
              type: string
              format: date-time
            end:
              type: string
              format: date-time
        totalRequests:
          type: integer
        totalTokens:
          type: integer
        totalCost:
          type: number
          format: float
        byProvider:
          type: object
          additionalProperties:
            type: object
            properties:
              requests:
                type: integer
              tokens:
                type: integer
              cost:
                type: number
                format: float
        byModel:
          type: object
          additionalProperties:
            type: object
            properties:
              requests:
                type: integer
              tokens:
                type: integer
              cost:
                type: number
                format: float
...
paths:
...
  /_api/projects:
    get:
      tags: [Projects]
      summary: List projects
      description: Get all projects the user has access to
      security:
        - FirebaseAuth: []
        - PersonalAccessToken: []
      responses:
        '200':
          description: Projects retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
                  total:
                    type: integer

    post:
      tags: [Projects]
      summary: Create project
      description: Create a new project
      security:
        - FirebaseAuth: []
        - PersonalAccessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Project created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
...
  /_api/projects/{id}/keys:
    get:
      tags: [API Keys]
      summary: List project API keys
      description: Get all API keys for a project (keys are masked for security)
      security:
        - FirebaseAuth: []
        - PersonalAccessToken: []
      parameters:
        - name: id
          in: path
          required: true
          description: Project ID
          schema:
            type: string
      responses:
        '200':
          description: API keys retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  keys:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApiKey'
                  total:
                    type: integer

    post:
      tags: [API Keys]
      summary: Add API key
      description: Add a new API key to the project (requires admin or owner role)
      security:
        - FirebaseAuth: []
        - PersonalAccessToken: []
      parameters:
        - name: id
          in: path
          required: true
          description: Project ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddApiKeyRequest'
      responses:
        '201':
          description: API key added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "API key added successfully"
                  keyId:
                    type: string
                  provider:
                    type: string
                  maskedKey:
                    type: string
                  addedAt:
                    type: string
                    format: date-time
...
  /_api/projects/{id}/usage:
    get:
      tags: [Analytics]
      summary: Get project usage statistics
      description: Get detailed usage statistics for a project
      security:
        - FirebaseAuth: []
        - PersonalAccessToken: []
      parameters:
        - name: id
          in: path
          required: true
          description: Project ID
          schema:
            type: string
        - name: startDate
          in: query
          description: Start date for usage period (ISO 8601)
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: End date for usage period (ISO 8601)
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Usage statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageStats'
...